To quicken up my workflow in Maya I tried to create automated actions through the Mel script. I learned to do this by opening the Mel script (entering ScriptEditor in the command line) doing the actions I wanted to automate, copying them into the command line then looking on the internet for syntax that would make the change relative. The only issue I faced was when copying the command over the line would call for that specific object, and so I found a way to make the code nonspecific and only effect the selected object.
90Degree.mel
Designed to rotate an object 90 degrees, avoiding the hassle of doing it rough and then entering it into the channel box.

FreezeTransformations;
rotate -r 1 90 1;

The code freezers the object to start with and then adds 90 into it's Y rotation channel.

WMD_X (World Mirror Duplicate X)
This code mirrors the object within the world, which would be very useful to creating symmetrical assets or proportions for games.

duplicate -rr;
doGroup 0 1 1;
scale -r -1 1 1 ;
ungroup;
FreezeTransformations;

The code duplicates the object, and then enters it into a group. With the group being orientated to the world origin, the code enters a -1 in scale X, mirroring it to the other side. The code then ungroups the object and freezes it's transforms.

WMD_Z (World Mirror Duplicate Z)
This is the same as the World Mirror X, but deals within the Z axis.

duplicate -rr;
doGroup 0 1 1;
scale -r 1 1 -1 ;
ungroup;
FreezeTransformations;